pyproject.toml
[tool.poetry]
name = "fastapi-blog"
version = "0.1.0"
description = ""
authors = ["Code of Mario <info@codeofmario.com>"]

[tool.poetry.dependencies]
python = ">3.10"
asyncio = "^3.4.3"
fastapi = {extras = ["all"], version = "^0.88.0"}
SQLAlchemy = "2.0.0b4"
minio = "^7.1.12"
python-multipart = "^0.0.5"
bcrypt = "^4.0.1"
PyJWT = "^2.6.0"
redis = "^5.0.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
